Renamed from Deoxide to Giggity.
Deoxide was derived from Oxegen, an Irish music festival which back in 2008
inspired me to make this program (since their paper schedule booklet was
worthless (and was sold separately for EUR 10 or so!!)). To avoid trademark
issues, I'm renaming it to Giggity. Even though I now primarily use this
for conferences, having "gig" in the name seems sensible.

Obviously I have *nothing* to do with Glenn Quagmire. :>