Block schedule font size update. OK, I give up. The formula made little sense, I'll admit that. Finding an optimal font size is not that simple IMHO, when you need to render schedules with silly-long names and/or silly-short slots.
But OK. Just default to showing ~3 lines in a block and offer a setting
for larger fonts or a static font size so people can get mostly the old
behaviour back. I hope this helps..

Also, figured out how to get the settings menu to show the current value
for each setting.