Ugh, id entry missing, I keep forgetting about that ~bug.
Bug because it's a PITA to deal with, even though I think I did have a
point back in 2008 when I didn't want the schedule URL to be *the*
unique primary key for a schedule. But that doesn't really help when
none of the relevant file formats have something that really counts as
an id field.