Improve barcode scanner error, adding an F-Droid link to the ZXing scanner since that seems to be the only way to still install it. (And actually the only scanner handling these intents, as I did not realise at the time.)
I'll try to update the ggt.gaa.st hack to support JSON imports so folks
can do that without zxing installed (just using any scanner or probably
the camera app). The export functionality however is I'm guessing
barely used so I'll drop that and/or hopefully replace it with a
sync backend.