Basic timezone awareness for Pentabarf-style schedules with the right info. The <date/> field will contain a full zoned ISO8601 timestamp. If available, it'll be used instead of simple naive <time/> and the offset to local gets remembered.
UI will then show the tz offset if there is one but otherwise will just show
all times as local.

This still interacts pretty poorly with the <day_change/> info.  I can
somewhat fix it but it will get confusing in cases where the change ends up
being 23:00 or something where it probably should be -1:00 which Giggity
does not currently support.

Let's leave that for later, for now I just want Giggity to be usable during
this COVID-19 period where attendance from home will obviously be the
standard for a while.