Start using `guid` attribute in Pentabarf/etc XML files instead of `id`, when available.
The latter seems to not actually be unique at all in some cases, while the former is not actually used in some others.

This should fix https://github.com/Wilm0r/giggity/issues/252, which BTW has a fun comment thread with a few examples of how this is a mess. :(