2.1.4.
Rather amusingly this is *exactly* a year after 2.1.3. Not much has changed TBH but there are some good bugfixes + Play Store wants the Android API bump.

While at it, remove some files that aren't actually used. It's an experimental adaptive icon that I ended up not using, though it is actually kind of clever and nice.